test(plugin-page-builder): cover drag activation across consecutive gestures
